// frontend/src/app/models/user.ts
export class User {
    password: string= '';
    firstname: string = '';
    lastname: string  = '';
    faculty: string= '';
    email: string= '';
    department: string= '';
    index: string=  '';
    year: number= 0;
    phone: string= '';
    type: string= '';
    country: string= '';
    accepted: boolean= false;
    registered: boolean= false;
    organizator: boolean= false;
    room: string= '';
    package:number=0;
}